Rename test.
